+++ passkey-types/src/ctap2/make_credential.rs
@@ -9,6 +9,8 @@ use crate::webauthn::{
     CollectedClientData, PublicKeyCredentialCreationOptions, PublicKeyCredentialDescriptor,
 };
 
+use super::extensions::{AuthenticatorPrfInputs, AuthenticatorPrfMakeOutputs, HmacGetSecretInput};
+
 serde_workaround! {
     /// While similar in structure to [`PublicKeyCredentialCreationOptions`],
     /// it is not completely identical, namely the presence of the `options` key.
@@ -201,7 +203,53 @@ const fn default_true() -> bool {
 
 /// All supported Authenticator extensions inputs during credential creation
 #[derive(Debug, Serialize, Deserialize, Default)]
-pub struct ExtensionInputs {}
+pub struct ExtensionInputs {
+    /// A boolean value to indicate that this extension is requested by the Relying Party
+    ///
+    /// <https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-errata-20220621.html#sctn-hmac-secret-extension>
+    #[serde(
+        rename = "hmac-secret",
+        default,
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub hmac_secret: Option<bool>,
+
+    /// The input salts for fetching and deriving a symmetric secret during registration.
+    ///
+    /// TODO: link to the hmac-secret-mc extension in the spec once it's published.
+    #[serde(
+        rename = "hmac-secret-mc",
+        default,
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub hmac_secret_mc: Option<HmacGetSecretInput>,
+
+    /// The direct input from a on-system client for the prf extension.
+    ///
+    /// The output from a request using the `prf` extension will not be signed
+    /// and will be un-encrypted.
+    /// This input should already be hashed by the client.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub prf: Option<AuthenticatorPrfInputs>,
+}
+
+impl ExtensionInputs {
+    /// Validates that there is at least one extension field that is `Some`.
+    /// If all fields are `None` then this returns `None` as well.
+    pub fn zip_contents(self) -> Option<Self> {
+        let Self {
+            hmac_secret,
+            hmac_secret_mc,
+            prf,
+        } = &self;
+
+        let has_hmac_secret = hmac_secret.is_some();
+        let has_hmac_secret_mc = hmac_secret_mc.is_some();
+        let has_prf = prf.is_some();
+
+        (has_hmac_secret || has_hmac_secret_mc || has_prf).then_some(self)
+    }
+}
 
 serde_workaround! {
     /// Upon successful creation of a credential, the authenticator returns an attestation object.
@@ -222,5 +270,61 @@ serde_workaround! {
         // the keys
         #[serde(rename = 0x03)]
         pub att_stmt: ciborium::value::Value,
+
+        /// A map, keyed by extension identifiers, to unsigned outputs of extensions, if any.
+        /// Authenticators SHOULD omit this field if no processed extensions define unsigned outputs.
+        /// Clients MUST treat an empty map the same as an omitted field.
+        #[serde(rename = 0x06, default, skip_serializing_if = Option::is_none)]
+        pub unsigned_extension_outputs: Option<UnsignedExtensionOutputs>,
+    }
+}
+
+/// All supported Authenticator extensions outputs during credential creation
+///
+/// This is to be serialized to [`Value`] in [`AuthenticatorData::extensions`]
+#[derive(Debug, Serialize, Deserialize)]
+pub struct SignedExtensionOutputs {
+    /// A boolean value to indicate that this extension was successfully processed by the extension
+    ///
+    /// <https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-errata-20220621.html#sctn-hmac-secret-extension>
+    #[serde(
+        rename = "hmac-secret",
+        default,
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub hmac_secret: Option<bool>,
+}
+
+impl SignedExtensionOutputs {
+    /// Validates that there is at least one extension field that is `Some`.
+    /// If all fields are `None` then this returns `None` as well.
+    pub fn zip_contents(self) -> Option<Self> {
+        let Self { hmac_secret } = &self;
+        let has_hmac_secret = hmac_secret.is_some();
+
+        (has_hmac_secret).then_some(self)
+    }
+}
+
+/// A map, keyed by extension identifiers, to unsigned outputs of extensions, if any.
+/// Authenticators SHOULD omit this field if no processed extensions define unsigned outputs.
+/// Clients MUST treat an empty map the same as an omitted field.
+#[derive(Debug, Serialize, Deserialize, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct UnsignedExtensionOutputs {
+    /// This output is supported in the Webauthn specification and will be used when the authenticator
+    /// and the client are in memory or communicating through an internal channel.
+    ///
+    /// If you are using transports where this needs to pass through a wire, use hmac-secret instead.
+    pub prf: Option<AuthenticatorPrfMakeOutputs>,
+}
+
+impl UnsignedExtensionOutputs {
+    /// Validates that there is at least one extension field that is `Some`.
+    /// If all fields are `None` then this returns `None` as well.
+    pub fn zip_contents(self) -> Option<Self> {
+        let Self { prf } = &self;
+
+        prf.is_some().then_some(self)
     }
 }